#2 Рефакторинг и новые компоненты
